keep only the best approach to tw refactor; combine categorical classes into one file
