+++ client/tw/test/TwRouter.integration.spec.ts
@@ -175,92 +175,3 @@ tape('extended TwBase', async test => {
 
 	test.end()
 })
-
-tape('handler by class', async test => {
-	if (!twlst) twlst = await getTws()
-
-	const msg = 'should convert handler instances to the extended interface'
-	try {
-		const data = {
-			sample1: { sex: 1, diaggrp: 'ALL' },
-			sample2: { sex: 2, diaggrp: 'NBL' }
-		}
-		//const xtws = terms.map(getHandler)
-		const start = Date.now()
-		const app = new FakeApp({ twlst, vocabApi, mode: 'handler' })
-		app.main(data)
-		test.pass(msg)
-		if (twlst.length > 10) {
-			// indicates benchmark test
-			console.log(195, `handler time, twlst.length=${twlst.length}`, Date.now() - start)
-			test.end()
-			return
-		}
-
-		const Inner = app.getInner()
-		test.deepEqual(
-			Object.keys(Inner.xtws[0]).sort(),
-			Object.keys(Inner.xtws[1]).sort(),
-			`should have matching handler property/method names for all extended handler instances`
-		)
-
-		test.equal(
-			Inner.dom.svg,
-			`<svg>` +
-				`<text>sample1, Male</text><circle r=1></cicle>` +
-				`<text>sample2, Female</text><circle r=2></cicle>` +
-				`<text>sample1, ALL</text><rect width=10 height=10></rect>` +
-				`<text>sample2, NBL</text><rect width=10 height=10></rect>` +
-				`</svg>`,
-			`should render an svg with fake data`
-		)
-	} catch (e) {
-		test.fail(msg + ': ' + e)
-	}
-
-	test.end()
-})
-
-tape('handler with addons', async test => {
-	if (!twlst) twlst = await getTws()
-
-	const msg = 'should convert handler instances to the extended interface'
-	try {
-		const data = {
-			sample1: { sex: 1, diaggrp: 'ALL' },
-			sample2: { sex: 2, diaggrp: 'NBL' }
-		}
-		const start = Date.now()
-		const app = new FakeApp({ twlst, vocabApi, mode: 'addons' })
-		app.main(data)
-		test.pass(msg)
-		if (twlst.length > 10) {
-			// indicates benchmark test
-			console.log(240, `addons time, twlst.length=${twlst.length}`, Date.now() - start)
-			test.end()
-			return
-		}
-
-		const Inner = app.getInner()
-		test.deepEqual(
-			Object.keys(Inner.xtws[0]).sort(),
-			Object.keys(Inner.xtws[1]).sort(),
-			`should have matching handler property/method names for all extended handler instances`
-		)
-
-		test.equal(
-			Inner.dom.svg,
-			`<svg>` +
-				`<text>sample1, Male</text><circle r=1></cicle>` +
-				`<text>sample2, Female</text><circle r=2></cicle>` +
-				`<text>sample1, ALL</text><rect width=10 height=10></rect>` +
-				`<text>sample2, NBL</text><rect width=10 height=10></rect>` +
-				`</svg>`,
-			`should render an svg with fake data`
-		)
-	} catch (e) {
-		test.fail(msg + ': ' + e)
-	}
-
-	test.end()
-})
