+++ client/tw/test/TwRouter.unit.spec.ts
@@ -5,6 +5,7 @@ import { vocabInit } from '#termdb/vocabulary'
 import { getExample } from '#termdb/test/vocabData'
 import { termjson } from '../../test/testdata/termjson'
 import { CategoricalRouter } from '../CategoricalRouter'
+import { CatValues, CatPredefinedGS, CatCustomGS } from '../categorical'
 
 const vocabApi = vocabInit({ state: { vocab: { genome: 'hg38-test', dslabel: 'TermdbTest' } } })
 
@@ -110,9 +111,8 @@ tape('initRaw() categorical', async test => {
 		}
 
 		const handler = await TwRouter.initRaw(tw, { vocabApi })
-		test.equal(
-			handler.router,
-			CategoricalRouter,
+		test.true(
+			handler instanceof CatValues,
 			`should return a matching categorical handler.router on init() with missing q or q.type`
 		)
 	}
@@ -125,9 +125,8 @@ tape('initRaw() categorical', async test => {
 		}
 
 		const handler = await TwRouter.initRaw(tw, { vocabApi })
-		test.equal(
-			handler.router,
-			CategoricalRouter,
+		test.true(
+			handler instanceof CatPredefinedGS,
 			`should return a matching categorical handler.router on init() with q.type='predefined-groupset'`
 		)
 	}
@@ -145,9 +144,8 @@ tape('initRaw() categorical', async test => {
 		}
 
 		const handler = await TwRouter.initRaw(tw, { vocabApi })
-		test.equal(
-			handler.router,
-			CategoricalRouter,
+		test.true(
+			handler instanceof CatCustomGS,
 			`should return a matching categorical handler.router on init() with q.type='custom-groupset'`
 		)
 	}
