--- client/tw/test/fake/extended/xCatTWPredefinedGS.ts
@@ -1,76 +1,0 @@
-import {
-	CategoricalTerm,
-	CatTWPredefinedGS,
-	CatTWCustomGS,
-	PredefinedGroupSettingQ,
-	BaseGroupSet,
-	ValuesQ
-} from '#types'
-import { HandlerOpts } from '../../../Handler'
-import { PlotTwRenderOpts } from '../types'
-import { TwBase } from '../../../TwBase'
-
-export class xCatTWPredefinedGS extends TwBase {
-	term: CategoricalTerm
-	q: PredefinedGroupSettingQ
-	#groupset: BaseGroupSet
-	#tw: CatTWPredefinedGS
-	#opts: HandlerOpts
-
-	// declare a constructor, to narrow the tw type
-	constructor(tw: CatTWPredefinedGS, opts: HandlerOpts = {}) {
-		super(tw, opts)
-		this.term = tw.term // to narrow to categorical term, since TwBase.term is just Term
-		this.q = tw.q
-		this.#tw = tw
-		this.#groupset = this.term.groupsetting[this.#tw.q.predefined_groupset_idx]
-		this.#opts = opts
-	}
-
-	render(arg: PlotTwRenderOpts) {
-		// the tw is guaranteed to have term.type=categorical, q.type='predefined-groupset'
-		const t = this.term
-		for (const [sampleId, d] of Object.entries(arg.data)) {
-			const keys = Object.keys(d)
-			// lots of terms indicate benchmark testing, must not by influenced by string-based svg simulated render
-			if (keys.length > 10 || !keys.includes(t.id)) continue
-			// for the tw in this typed context, use a svg:rect element
-			const shape = `<rect width=10 height=10></rect></svg>`
-			arg.holder = arg.holder.replace(`</svg>`, `<text>${sampleId}, ${d[t.id]}</text>${shape}`)
-		}
-	}
-}
-
-//test only
-const tw = {
-	type: 'CatTWPredefinedGS' as const,
-	term: {
-		type: 'categorical' as const,
-		id: 'abc',
-		name: 'ABC',
-		values: {
-			x: { label: 'x' },
-			y: { label: 'y' }
-		},
-		groupsetting: {
-			lst: [
-				{
-					name: 'test',
-					groups: []
-				}
-			]
-		}
-	},
-	q: {
-		type: 'predefined-groupset' as const,
-		predefined_groupset_idx: 0
-	}
-	// q: {
-	// 	type: 'custom-groupset' as const,
-	// 	customset: {
-	// 		groups: []
-	// 	}
-	// },
-}
-
-const a = new xCatTWPredefinedGS(tw)
